+++ previous-versions/banking/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [python v0.28.0] previous-versions/banking
 ### Releases
 - [PyPI v0.28.0] https://pypi.org/project/codat-banking/0.28.0 - previous-versions/banking
+
+## 2023-11-27 01:30:17
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Banking.yaml
+- Speakeasy CLI 1.122.0 (2.195.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.29.0] previous-versions/banking
+### Releases
+- [PyPI v0.29.0] https://pypi.org/project/codat-banking/0.29.0 - previous-versions/banking